feat: redis 직렬화를 위한 inner class 분리
