+++ usecase/src/main/java/org/depromeet/spot/usecase/port/in/review/BlockKeywordInfo.java
@@ -1,0 +1,14 @@
+package org.depromeet.spot.usecase.port.in.review;
+
+import java.io.Serializable;
+
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+
+import lombok.Builder;
+
+@JsonSerialize
+@JsonDeserialize
+@Builder
+public record BlockKeywordInfo(String content, Long count, Boolean isPositive)
+        implements Serializable {}
