+++ usecase/src/main/java/org/depromeet/spot/usecase/port/in/review/LocationInfo.java
@@ -1,0 +1,9 @@
+package org.depromeet.spot.usecase.port.in.review;
+
+import java.io.Serializable;
+
+import lombok.Builder;
+
+@Builder
+public record LocationInfo(String stadiumName, String sectionName, String blockCode)
+        implements Serializable {}
