+++ usecase/src/main/java/org/depromeet/spot/usecase/port/in/review/MyRecentReviewResult.java
@@ -1,0 +1,8 @@
+package org.depromeet.spot.usecase.port.in.review;
+
+import org.depromeet.spot.domain.review.Review;
+
+import lombok.Builder;
+
+@Builder
+public record MyRecentReviewResult(Review review, Long reviewCount) {}
